zmart-api/services/gpt-mds-agent/requirements.txt
# Core dependencies
openai==1.12.0
aiohttp==3.9.1
pyyaml==6.0.1
jinja2==3.1.3
python-dotenv==1.0.0
pydantic==2.5.0
asyncio==3.4.3

# Database & caching
aiosqlite==0.19.0
redis==5.0.1
cachetools==5.3.2

# Monitoring & logging
prometheus-client==0.19.0
structlog==23.2.0
sentry-sdk==1.39.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
aioresponses==0.7.6

# Development tools
black==23.12.0
flake8==6.1.0
mypy==1.7.1
ipython==8.18.1

# Additional utilities
backoff==2.2.1
aiohttp-cors==0.7.0
